ConvAI_Assignment2/requirements.txt
accelerate==0.33.0
faiss-cpu==1.8.0.post1
matplotlib==3.8.4   # downgraded for Windows compatibility
nltk==3.9.1
numpy==1.26.4
pandas==2.2.2
peft==0.11.1
pymupdf==1.24.10
python-dotenv==1.0.1
rank-bm25==0.2.2
regex==2024.5.15
reportlab==4.2.2
scikit-learn==1.5.1
sentence-transformers==2.7.0
streamlit==1.37.1
torch>=2.2,<2.6
transformers==4.44.2
tqdm==4.66.4
datasets==2.20.0